3d pages init
Made the 3d pages so I could get started on those. Lots of work to do, notably the fact that the background resizing function doesn't work quite like I want it to because the background on the 3d pages is positioned to the right. Bummer! Thinking I'll have to have to implement a function that measures the hypothetical scrollbar of a page and subtract that from the 3d background position to make up for the change in position that the scrollbar causes.