Transitions working on Firefox
Transitions into the index page are mostly working (with some hiccups)... but not at all on Chrome. It's not my personal browser but IS a pretty popular one, so getting things working on there is really important. The issue is that Chrome doesn't seem to think with my js as quick as Firefox does, so I need to come up with some way to make the browser "remember" that the index page needs to load a specific way when it's travelled to from one of the info pages.
The hiccups in my Firefox version is the fact that transitions into the index don't work if you start on a non-index page; again, I need to come up with some way for the browser to know that if it's coming into the index from an info page, there are elements in that index that need certain classes to animate.
Also, back to wrestling with fixBackgroundSizeCover. It seems like the listeners get caught up in whatever page they were first loaded into (I imagine this is due to AJAX), and therefore make the console pretty ugly when you end up on a page that doesn't have any elements with an id that it would've been expecting to see on a previous page. For that, I need to come up with some way to make it so that the fixBackgroundCover function doesn't panic when it's value doesn't exist.