Animations into index running (sort of)
What it says on the tin. Running into an issue with the footer not being positioned absolute on the bottom of info/form pages; it pops into existence when moving into the index page and I don't like it. Also as a result of the footer not being absolutely positioned, the flexboxes for the backgrounds don't meet in the middle of their intended container unless I subtract the height of the footer from their 100vh height, which leads to the flexbox transition overshooting it's animation when coming into the index. Might be able to save this by making the height for one of the backgrounds on the index shorter/longer than the other.