Logo animations running
Got to the bottom of my previous issue. Easy solution of course; just needed to give the logo elements heights of 100%, yayyyy.
I need to make new crops of the backgrounds and get rid of those bouts of inline styles, then I'll be pretty much done with the mobile layout stuff. Exciting! I simultaneously have very little and a ton to do.