File upload more secure
Was trying to figure out form ajax silliness, got distracted by the possible lack of security in file uploading. The js is a lot stricter with it now and does not allow form submission if the file upload gives any errors until the user redoes their upload with new, acceptable files.