Submission form running, can relay errors
The forms are complete. They send emails with attachments and, if form isn't filled out correctly, relay to the user why the form won't send so that they can fix it.
I also added a modal  to appear when the form is sending. I noticed in my testing that image uploads can take a while if the file is big, so I want to be sure the user knows they need to wait and to not close/refresh the page.